requirements.txt
#
# Modifications © 2019 Hashmap, Inc
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

-i https://pypi.org/simple
alembic==0.9.10
apache-airflow==1.10.3
babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
click==7.0
colorama==0.4.1
configparser==3.5.3
croniter==0.3.30
defusedxml==0.6.0
dill==0.2.9
docutils==0.14
flask-admin==1.5.3
flask-appbuilder==1.12.3
flask-babel==0.12.2
flask-caching==1.3.3
flask-login==0.4.1
flask-openid==1.2.5
flask-sqlalchemy==2.4.0
flask-swagger==0.2.13
flask-wtf==0.14.2
flask==1.0.3
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
gitpython==2.1.11
gunicorn==19.9.0
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
jinja2==2.10
json-merge-patch==0.2
lockfile==0.12.2
lxml==4.3.4
mako==1.0.13
markdown==2.6.11
markupsafe==1.1.1
numpy==1.16.4
ordereddict==1.1
pandas==0.24.2
pendulum==1.4.4
psutil==5.6.3
pygments==2.4.2
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.2
pyyaml==5.1.1
requests==2.22.0
setproctitle==1.1.10
six==1.12.0
smmap2==2.0.5
sqlalchemy==1.2.19
tabulate==0.8.3
tenacity==4.12.0
text-unidecode==1.2
thrift==0.11.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
werkzeug==0.14.1
wtforms==2.2.1
zope.deprecation==4.4.0
